# src/swaggerYaml/swagger.yaml
openapi: 3.0.0
info:
  title: Appointment Booking API
  description: |
    API for managing appointments with role-based access control.
    Real-time notifications via Socket.io.
  version: 1.0.0

servers:
  - url: https://appointment-booking-ghjp.onrender.com
    description: "Main Server (Render)"
  - url: http://localhost:3000
    description: "Local Development"

security:
  - bearerAuth: []

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                role: { type: string, enum: [client, provider] }
      responses:
        "201": { description: Created }
        "400": { description: Validation error or user exists }

  /api/auth/login:
    post:
      summary: Login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: Success }
        "400": { description: Invalid credentials }

  /api/providers:
    post:
      summary: Onboard as a provider
      tags: [Providers]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                specialty: { type: string }
                bio: { type: string }
      responses:
        "201": { description: Provider profile created }
        "400": { description: Already a provider }
    get:
      summary: List all providers
      tags: [Providers]
      responses:
        "200": { description: Array of providers }

  /api/slots:
    post:
      summary: Create a slot (Provider only)
      tags: [Slots]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startTime, endTime]
              properties:
                startTime: { type: string, format: date-time }
                endTime: { type: string, format: date-time }
      responses:
        "201": { description: Created }
        "403": { description: Not a provider }
    get:
      summary: List available slots
      tags: [Slots]
      responses:
        "200": { description: Array of available slots }

  /api/slots/{id}:
    get:
      summary: Get a slot by ID
      tags: [Slots]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Slot object }
        "404": { description: Not found }

  /api/slots/{id}/book:
    patch:
      summary: Book a slot
      tags: [Slots]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Booked }
        "400": { description: Already booked }
        "404": { description: Not found }

  /api/slots/{id}/cancel:
    patch:
      summary: Cancel a booked slot
      tags: [Slots]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Cancelled }
        "400": { description: Not booked }
        "403": { description: Not authorized }
        "404": { description: Not found }

  /api/appointments:
    get:
      summary: Get appointments for logged-in user
      tags: [Appointments]
      security:
        - bearerAuth: []
      responses:
        "200": { description: Array of appointments }

  /api/appointments/book:
    post:
      summary: Book an appointment
      tags: [Appointments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slot_id: { type: string }
      responses:
        "201": { description: Booked }
        "400": { description: Already booked }

  /api/appointments/cancel:
    post:
      summary: Cancel an appointment
      tags: [Appointments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slot_id: { type: string }
      responses:
        "200": { description: Cancelled }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT